Bugfix: Do not run pump after all channels are "CHEM_WASH"-ed
The `break;` statement after the `CHEM_WASH` loop was leading to the
code falling through to the statements that would turn on the
pump.

The problem was, after the loop, all the valves are reset. This closes
all the input ports, therefore, the pump runs with an invalid valve
setup. A valid valve setup is such that there is an open path from one
input port (a chemical channel or wash or air-vent) to one output
port (waste or collection).

This commit ensures that the function exits (returns) immediately
after the loop is complete.